+++ R/er_farm_aim.R
@@ -78,9 +78,9 @@ er_farm_aim <- function(B_SOILTYPE_AGR, B_AREA, medalscore = "gold", farmscore =
   out.tgt <- dcast(out.tgt,medalscores~indicator,value.var = 'er_score')
   
   # add target costs and total farm score on farm level
-  out.tgt[medalscore == "gold", c('B_CT_FARM_TOT','B_CT_COSTS') := list(35,175)]
+  out.tgt[medalscore == "gold", c('B_CT_FARM_TOT','B_CT_COSTS') := list(35,200)]
   out.tgt[medalscore == "silver", c('B_CT_FARM_TOT','B_CT_COSTS') := list(22,100)]
-  out.tgt[medalscore == "bronze", c('B_CT_FARM_TOT','B_CT_COSTS') := list(14,70)]
+  out.tgt[medalscore == "bronze", c('B_CT_FARM_TOT','B_CT_COSTS') := list(14,60)]
   
   # add a farm id
   out.tgt[,farmid := 1]
@@ -111,24 +111,24 @@ er_farm_aim <- function(B_SOILTYPE_AGR, B_AREA, medalscore = "gold", farmscore =
     out.threshold[medalscores == "gold",s_er_farmtotal := 35]
     
     # add target costs on farm level
-    out.threshold[medalscores == "gold", s_er_costs := (175/175)*100]
-    out.threshold[medalscores == "silver", s_er_costs := (110/175)*100]
-    out.threshold[medalscores == "bronze", s_er_costs := (70/175)*100]
+    out.threshold[medalscores == "gold", s_er_costs := (200/200)*100]
+    out.threshold[medalscores == "silver", s_er_costs := (100/200)*100]
+    out.threshold[medalscores == "bronze", s_er_costs := (60/200)*100]
     
     # round values 
     cols <- colnames(out.threshold)[grepl('er_costs',colnames(out.threshold))]
     out.threshold[,c(cols) := lapply(.SD,round,1),.SDcols = cols]
     
-    # set threshold of golden medal for landscape to 0.5 and
+    # set threshold of golden medal for landscape to 0 and
     # remove thresholds of bronze and silver medal for landscape
-    out.threshold[medalscores == "gold", cf_landscape := 0.5]
+    out.threshold[medalscores == "gold", cf_landscape := 0]
     out.threshold[medalscores == "silver", cf_landscape := NA_real_]
     out.threshold[medalscores == "bronze", cf_landscape := NA_real_]
  
     # if farm only includes peat soils, set threshold for water to 0.5 
     if( all(grepl("veen",B_SOILTYPE_AGR)) == TRUE){
     
-    out.threshold[medalscores == "gold", cf_water := 0.5]
+    out.threshold[medalscores == "gold", cf_water := 0]
     out.threshold[medalscores == "silver", cf_water := NA_real_]
     out.threshold[medalscores == "bronze", cf_water := NA_real_]
     }