+++ R/er_main.R
@@ -88,9 +88,9 @@ ecoregeling <- function(B_SOILTYPE_AGR, B_LU_BRP,B_LU_BBWP,
                               B_AREA = B_AREA, type = 'farm')]
   
   # correct total reward in dt.farm after medal is awarded
-  dt.farm[medal == "bronze", S_ER_REWARD := 70]
-  dt.farm[medal == "silver", S_ER_REWARD := 110]
-  dt.farm[medal == "gold", S_ER_REWARD := 175]
+  dt.farm[medal == "bronze", S_ER_REWARD := 60]
+  dt.farm[medal == "silver", S_ER_REWARD := 100]
+  dt.farm[medal == "gold", S_ER_REWARD := 200]
   dt.farm[medal == "none", S_ER_REWARD := 0]
 
   # estimate the opportunity index for farm and field
@@ -191,16 +191,16 @@ ecoregeling <- function(B_SOILTYPE_AGR, B_LU_BRP,B_LU_BBWP,
     out.farm[, s_er_reward := dt.farm$S_ER_REWARD]
     out.farm[, s_er_tot := dt.opi$dt.farm.score]
     
-    # set maximum for s_er_costs at 175 and convert to percentage 
-    out.farm[, s_er_costs := (pmin(175,s_er_costs)/175)*100]
+    # set maximum for s_er_costs at 200 and convert to percentage 
+    out.farm[, s_er_costs := (pmin(200,s_er_costs)/200)*100]
     
     # set maximum for eco scores and total farm scores on farm level
-    out.farm[, s_er_soil := pmin(15,s_er_soil)]
-    out.farm[, s_er_water := pmin(15,s_er_water)]
-    out.farm[, s_er_climate := pmin(15,s_er_climate)]
-    out.farm[, s_er_biodiversity := pmin(15,s_er_biodiversity)]
+    out.farm[, s_er_soil := pmin(dt.farm.thresholds$s_er_soil_gold,s_er_soil)] 
+    out.farm[, s_er_water := pmin(dt.farm.thresholds$s_er_water_gold,s_er_water)]
+    out.farm[, s_er_climate := pmin(dt.farm.thresholds$s_er_climate_gold,s_er_climate)]
+    out.farm[, s_er_biodiversity := pmin(dt.farm.thresholds$s_er_biodiversity_gold,s_er_biodiversity)]
     out.farm[, s_er_landscape := pmin(1,s_er_landscape)]
-    out.farm[, s_er_farm_tot:= pmin(50,s_er_farm_tot)]
+    out.farm[, s_er_farm_tot:= pmin(dt.farm.thresholds$s_er_farmtotal_gold,s_er_farm_tot)]
     
     # add thresholds
     out <- list(farm = c(as.list(out.farm),dt.farm.thresholds),